some changes

+++ db/sqlc/account.sql.go
@@ -21,7 +21,7 @@ INSERT INTO accounts (
 
 type CreateAccountParams struct {
 	Owner    string `json:"owner"`
-	Balance  int32  `json:"balance"`
+	Balance  int64  `json:"balance"`
 	Currency string `json:"currency"`
 }
 
@@ -116,7 +116,7 @@ RETURNING id, owner, balance, currency, created_at
 
 type UpdateAccountParams struct {
 	ID      int64 `json:"id"`
-	Balance int32 `json:"balance"`
+	Balance int64 `json:"balance"`
 }
 
 func (q *Queries) UpdateAccount(ctx context.Context, arg UpdateAccountParams) (Account, error) {
